requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.58.1
packaging==24.2
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
